ssd/requirements.txt
tensorflow-gpu==1.4.0
matplotlib==2.1.1
pillow==4.3.0
